backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12

# Database & ORM
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35

# Supabase
supabase==2.9.1
python-jose[cryptography]==3.3.0

# Redis & Celery
redis==5.1.1
celery==5.4.0

# LangChain & LangGraph
langchain==0.3.6
langchain-openai==0.2.3
langchain-community==0.3.4
langgraph==0.2.39

# Web Search
tavily-python==0.5.0

# Vector Store
chromadb==0.5.12

# Document Processing
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.1.1
httpx==0.27.2
python-dotenv==1.0.1
